+++ src/test/suite/helper/common.test.ts
@@ -16,6 +16,7 @@ import {
   getUnitUid,
   updateUnitUid,
   replaceUnitPatternPlaceholder,
+  formatModuleName
 } from "../../../helper/common";
 
 const expect = chai.expect;
@@ -161,5 +162,14 @@ suite("Common", () => {
     );
     replaceUnitPatternPlaceholder(filePath, "excercise");
     expect(spy).to.be.have.been.called;
+  });
+  test("Format module name", async () => {
+    const filePath = resolve(
+      __dirname,
+      "../../../../src/test/data/repo/learn-scaffolding-main/terms.json"
+    );
+    const moduleName = "Choose the best Service For Your Application";
+    const moduleFolderName = formatModuleName(moduleName, filePath);
+    expect(moduleFolderName).to.equal("choose-best-service-for-your-app");
   });
 });